+++ source/state_representation/test/tests/test_trajectory.cpp
@@ -1,5 +1,6 @@
 #include <chrono>
 #include <gtest/gtest.h>
+#include <state_representation/space/cartesian/CartesianState.hpp>
 #include <stdexcept>
 #include <type_traits>
 
@@ -114,26 +115,44 @@ TEST(TrajectoryTest, ConstructTrajectory) {
   EXPECT_NO_THROW(CartesianTrajectory trajectory("foo"));
   EXPECT_NO_THROW(CartesianTrajectory trajectory("foo", CartesianState::Random("foo"), std::chrono::nanoseconds(100)));
   EXPECT_NO_THROW(
-      CartesianTrajectory trajectory("foo", {CartesianState::Random("foo")}, {std::chrono::nanoseconds(100)})
-  );
+      CartesianTrajectory trajectory("foo", {CartesianState::Random("foo")}, {std::chrono::nanoseconds(100)}));
 
   EXPECT_THROW(
       CartesianTrajectory trajectory("foo", CartesianState(), std::chrono::nanoseconds(100)),
-      exceptions::EmptyStateException
-  );
+      exceptions::EmptyStateException);
   EXPECT_THROW(
       CartesianTrajectory trajectory(
           "foo", {CartesianState::Random("foo", "some_world"), CartesianState::Random("foo")},
-          {std::chrono::nanoseconds(100), std::chrono::nanoseconds(200)}
-      ),
-      exceptions::IncompatibleReferenceFramesException
-  );
+          {std::chrono::nanoseconds(100), std::chrono::nanoseconds(200)}),
+      exceptions::IncompatibleReferenceFramesException);
   EXPECT_THROW(
       CartesianTrajectory trajectory(
-          "foo", {CartesianState::Random("foo"), CartesianState::Random("foo")}, {std::chrono::nanoseconds(100)}
-      ),
-      exceptions::IncompatibleSizeException
-  );
+          "foo", {CartesianState::Random("foo"), CartesianState::Random("foo")}, {std::chrono::nanoseconds(100)}),
+      exceptions::IncompatibleSizeException);
+
+  CartesianTrajectory ct(
+      "foo", {CartesianState::Random("foo"), CartesianState::Random("bar"), CartesianState::Random("baz")},
+      {std::chrono::nanoseconds(100), std::chrono::nanoseconds(200), std::chrono::nanoseconds(300)});
+  EXPECT_NO_THROW(CartesianTrajectory ct_copy(ct));
+  {
+    CartesianTrajectory ct_copy(ct);
+    EXPECT_EQ(ct.get_size(), ct_copy.get_size());
+    for (unsigned int i = 0; i < ct.get_size(); ++i) {
+      EXPECT_EQ(ct.get_point(i).data(), ct_copy.get_point(i).data());
+      EXPECT_EQ(ct.get_duration(i), ct_copy.get_duration(i));
+    }
+    EXPECT_STREQ(ct.get_reference_frame().c_str(), ct_copy.get_reference_frame().c_str());
+  }
+  {
+    CartesianTrajectory ct_copy;
+    EXPECT_NO_THROW(ct_copy = ct);
+    EXPECT_EQ(ct.get_size(), ct_copy.get_size());
+    for (unsigned int i = 0; i < ct.get_size(); ++i) {
+      EXPECT_EQ(ct.get_point(i).data(), ct_copy.get_point(i).data());
+      EXPECT_EQ(ct.get_duration(i), ct_copy.get_duration(i));
+    }
+    EXPECT_STREQ(ct.get_reference_frame().c_str(), ct_copy.get_reference_frame().c_str());
+  }
 
   // Joint trajectory
   EXPECT_NO_THROW(JointTrajectory trajectory("foo"));
@@ -141,21 +160,40 @@ TEST(TrajectoryTest, ConstructTrajectory) {
   EXPECT_NO_THROW(JointTrajectory trajectory("foo", {JointState::Random("foo", 25)}, {std::chrono::nanoseconds(100)}));
 
   EXPECT_THROW(
-      JointTrajectory trajectory("foo", JointState(), std::chrono::nanoseconds(100)), exceptions::EmptyStateException
-  );
+      JointTrajectory trajectory("foo", JointState(), std::chrono::nanoseconds(100)), exceptions::EmptyStateException);
   EXPECT_THROW(
       JointTrajectory trajectory(
           "foo", {JointState::Random("foo", 25), JointState::Random("foo", 24)},
-          {std::chrono::nanoseconds(100), std::chrono::nanoseconds(200)}
-      ),
-      exceptions::IncompatibleStatesException
-  );
+          {std::chrono::nanoseconds(100), std::chrono::nanoseconds(200)}),
+      exceptions::IncompatibleStatesException);
   EXPECT_THROW(
       JointTrajectory trajectory(
-          "foo", {JointState::Random("foo", 25), JointState::Random("foo", 25)}, {std::chrono::nanoseconds(100)}
-      ),
-      exceptions::IncompatibleSizeException
-  );
+          "foo", {JointState::Random("foo", 25), JointState::Random("foo", 25)}, {std::chrono::nanoseconds(100)}),
+      exceptions::IncompatibleSizeException);
+
+  JointTrajectory jt(
+      "foo", {JointState::Random("foo", 25), JointState::Random("bar", 25), JointState::Random("baz", 25)},
+      {std::chrono::nanoseconds(100), std::chrono::nanoseconds(200), std::chrono::nanoseconds(300)});
+  EXPECT_NO_THROW(JointTrajectory ct_copy(jt));
+  {
+    JointTrajectory ct_copy(jt);
+    EXPECT_EQ(jt.get_size(), ct_copy.get_size());
+    for (unsigned int i = 0; i < jt.get_size(); ++i) {
+      EXPECT_EQ(jt.get_point(i).data(), ct_copy.get_point(i).data());
+      EXPECT_EQ(jt.get_duration(i), ct_copy.get_duration(i));
+    }
+    EXPECT_EQ(jt.get_joint_names(), ct_copy.get_joint_names());
+  }
+  {
+    JointTrajectory ct_copy;
+    EXPECT_NO_THROW(ct_copy = jt);
+    EXPECT_EQ(jt.get_size(), ct_copy.get_size());
+    for (unsigned int i = 0; i < jt.get_size(); ++i) {
+      EXPECT_EQ(jt.get_point(i).data(), ct_copy.get_point(i).data());
+      EXPECT_EQ(jt.get_duration(i), ct_copy.get_duration(i));
+    }
+    EXPECT_EQ(jt.get_joint_names(), ct_copy.get_joint_names());
+  }
 }
 
 TYPED_TEST_P(TrajectoryTest, AddRemovePoints) {
@@ -220,16 +258,14 @@ TYPED_TEST_P(TrajectoryTest, AddRemovePoints) {
   // additons and insertions of multiple points
   std::vector<PointType> points = {point0, point1, point2};
   std::vector<std::chrono::nanoseconds> durations = {
-      std::chrono::nanoseconds(10), std::chrono::nanoseconds(20), std::chrono::nanoseconds(30)
-  };
+      std::chrono::nanoseconds(10), std::chrono::nanoseconds(20), std::chrono::nanoseconds(30)};
   EXPECT_NO_THROW(this->add_points(points, durations));
   for (unsigned int i = 0; i < this->trajectory->get_size(); ++i) {
     this->expect_equal(points[i], i, durations[i]);
   }
   std::vector<PointType> shuffled_points = {point2, point0, point1};
   std::vector<std::chrono::nanoseconds> shuffled_durations = {
-      std::chrono::nanoseconds(30), std::chrono::nanoseconds(10), std::chrono::nanoseconds(20)
-  };
+      std::chrono::nanoseconds(30), std::chrono::nanoseconds(10), std::chrono::nanoseconds(20)};
   EXPECT_NO_THROW(this->set_points(shuffled_points, shuffled_durations));
   for (unsigned int i = 0; i < this->trajectory->get_size(); ++i) {
     this->expect_equal(shuffled_points[i], i, shuffled_durations[i]);
@@ -274,8 +310,7 @@ TYPED_TEST_P(TrajectoryTest, Exceptions) {
 
   std::vector<PointType> points = {point0, point1, point2};
   std::vector<std::chrono::nanoseconds> durations = {
-      std::chrono::nanoseconds(10), std::chrono::nanoseconds(20), std::chrono::nanoseconds(30)
-  };
+      std::chrono::nanoseconds(10), std::chrono::nanoseconds(20), std::chrono::nanoseconds(30)};
   EXPECT_NO_THROW(this->add_points(points, durations));
 
   EXPECT_THROW(this->delete_point_index(10), std::out_of_range);
@@ -338,8 +373,7 @@ TYPED_TEST_P(TrajectoryTest, Getters) {
 
   std::vector<PointType> points = {point0, point1, point2};
   std::vector<std::chrono::nanoseconds> durations = {
-      std::chrono::nanoseconds(10), std::chrono::nanoseconds(20), std::chrono::nanoseconds(30)
-  };
+      std::chrono::nanoseconds(10), std::chrono::nanoseconds(20), std::chrono::nanoseconds(30)};
   EXPECT_NO_THROW(this->add_points(points, durations));
   ASSERT_FALSE(this->trajectory->is_empty());
   auto trajectory_points = this->trajectory->get_points();